requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.0
whisper==1.1.10
elevenlabs==0.2.26
langchain==0.0.335
langchain-community==0.0.7
faiss-cpu==1.7.4
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pydub==0.25.1
numpy>=1.26
pandas<2.1
pypdf2==3.0.1
chromadb==0.4.15
speechrecognition==3.10.0
pyaudio==0.2.11
gtts==2.4.0
pygame==2.5.2
regex==2023.10.3
tiktoken==0.5.1